Refactored application startup and kafka boot to match other services. Changed kafka actor lifecycle to a local actor for now.
