+++ server/common/src/main/scala/com/eigengo/lift/common/MessagePayload.scala
@@ -1,4 +1,4 @@
-package com.eigengo.lift.kafka
+package com.eigengo.lift.common
 
 object MessagePayload {
   type Payload = Array[Byte]
