+++ server/main/src/main/scala/com/eigengo/lift/LiftLocalApp.scala
@@ -7,6 +7,7 @@ import akka.persistence.journal.leveldb.{SharedLeveldbJournal, SharedLeveldbStor
 import akka.util.Timeout
 import com.eigengo.lift.common.MicroserviceApp.MicroserviceProps
 import com.eigengo.lift.exercise._
+import com.eigengo.lift.kafka.{KafkaBoot, KafkaProducerActor}
 import com.eigengo.lift.notification.NotificationBoot
 import com.eigengo.lift.profile.ProfileBoot
 import com.typesafe.config.ConfigFactory
@@ -45,19 +46,11 @@ object LiftLocalApp extends App {
       startupSharedJournal(system, startStore = port == firstSeedNodePort, path = ActorPath.fromString(s"akka.tcp://$LiftActorSystem@127.0.0.1:$firstSeedNodePort/user/store"))
 
       // boot the microservices
-      //TODO: FIX
-      val manager = system.actorOf(ClusterSingletonManager.props(
-        KafkaProducerActor.props(config),
-        "kafka-producer",
-        "die",
-        None), "manager")
 
-      val kafka = system.actorOf(ClusterSingletonProxy.props(
-        "/user/manager/kafka-producer", None), "kafka-producer-proxy")
-
-      val profile = ProfileBoot.boot(system)
+      val kafka = KafkaBoot.boot(config)
+      val profile = ProfileBoot.boot
       val notification = NotificationBoot.boot
-      val exercise = ExerciseBoot.boot(kafka, notification.notification, profile.userProfile)
+      val exercise = ExerciseBoot.boot(kafka.kafka, notification.notification, profile.userProfile)
 
       startupHttpService(system, port, exercise.route(system.dispatcher), profile.route(system.dispatcher))
     }
